backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.1.0
pymongo==4.13.1
motor==3.7.1
firebase-admin==6.9.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.2.0
pytest
pytest-asyncio
httpx
mongomock-motor
pytest-env
pytest-cov
